requirements.txt
Django>=6.0
djangorestframework
psycopg2-binary
celery
redis
openpyxl